fixed errors

+++ src/Components/Orders/OrderFunctions.js
@@ -32,9 +32,9 @@ export const validateSubmit = (stateOrder,symbolsState,statusPair) => {
       return "You need choice two cryptos for the order";
     if(statusPair === 3)
       return "The pair is invalid"
-    if(Math.sign(parseInt(stateOrder.amount)) === -1 || Math.sign(parseInt(stateOrder.amount)) === 0)
+    if(Math.sign(parseFloat(stateOrder.amount)) === -1 || Math.sign(parseFloat(stateOrder.amount)) === 0)
       return "Amount cant be negative or zero"
-    if(stateOrder.type === "Limit" && (Math.sign(parseInt(stateOrder.limit)) === -1 || Math.sign(parseInt(stateOrder.limit)) === 0))
+    if(stateOrder.type === "Limit" && (Math.sign(parseFloat(stateOrder.limit)) === -1 || Math.sign(parseFloat(stateOrder.limit)) === 0))
       return "Limit cant be negative number or zero "
       return "";
 }
